+++ mangadex_library/lib/models/common/mangaReadingStatus.dart
@@ -1,0 +1,9 @@
+class MangaReadingStatus {
+  late String result;
+  late String status;
+  MangaReadingStatus(this.result, this.status);
+  MangaReadingStatus.fromJson(Map<String, dynamic> json) {
+    result = json['result']!;
+    status = json['status']!;
+  }
+}