+++ mangadex_library/lib/models/common/resultOk.dart
@@ -1,0 +1,15 @@
+class ResultOk {
+  late String result;
+
+  ResultOk({required this.result});
+
+  ResultOk.fromJson(Map<String, dynamic> json) {
+    result = json['result'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final data = <String, dynamic>{};
+    data['result'] = result;
+    return data;
+  }
+}