+++ mangadex_library/lib/models/user/logged_user_details/logged_user_details.dart
@@ -2,18 +2,14 @@ import 'package:mangadex_library/models/common/relationships.dart';
 
 class UserDetails {
   late final String result;
+  late final String response;
   late final Data data;
-  late final List<Relationship> relationships;
-  UserDetails(this.result, this.data, this.relationships);
+
+  UserDetails(this.result, this.data);
   UserDetails.fromJson(Map<String, dynamic> json) {
-    result = json['result'];
-    data = Data.fromjson(json['data']);
-    relationships = <Relationship>[];
-    if (json['relationships'] != null) {
-      json['relationships'].forEach((v) {
-        relationships.add(Relationship.fromJson(v));
-      });
-    }
+    result = json['result'] ?? '';
+    response = json['response'] ?? '';
+    data = Data.fromjson(json['data']!);
   }
 }
 
@@ -21,35 +17,30 @@ class Data {
   late final String id;
   late final String type;
   late final Attributes attributes;
-  Data(this.id, this.type, this.attributes);
+  late final List<Relationship> relationships;
+  Data(this.id, this.type, this.attributes, this.relationships);
   Data.fromjson(Map<String, dynamic> json) {
     id = json['id'] ?? '';
     type = json['type'] ?? '';
-    attributes = Attributes.fromJson(json['attributes']);
+    attributes = Attributes.fromJson(json['attributes']!);
+    relationships = [];
+    json['relationships']!.forEach((v) {
+      relationships.add(Relationship.fromJson(v));
+    });
   }
 }
 
 class Attributes {
   late final String username;
-  late final Role roles;
+  late final List<String> roles;
   late final int version;
   Attributes(this.username, this.roles, this.version);
   Attributes.fromJson(Map<String, dynamic> json) {
-    username = json['username'];
-    roles = Role.fromJson(json['roles']);
+    username = json['username'] ?? '';
+    roles = [];
+    json['roles'].forEach((v) {
+      roles.add(v);
+    });
     version = json['version'] ?? 0;
-  }
-}
-
-class Role {
-  late final List<String> roles;
-  Role(this.roles);
-  Role.fromJson(Map<String, dynamic> json) {
-    roles = <String>[];
-    if (json['roles'] != null) {
-      json['roles'].forEach((v) {
-        roles.add(v);
-      });
-    }
   }
 }